epio_launcher will load secret key from bundle config
I noticed that bundle_config.config['core']['secret_key'] was a secret
key, and that it persisted across versions. So, I added support for
autoloading it into the config. It can be turned off in the config
file, so that if you want to manage your own secrets you can.